Add basic working cosa example
